address comments from cran submission process
The following issues identified by Gregor Seyer were addressed in this resubmission:

> Please omit the redundant "R interface" in your description.

Redudancy removed from the description.

> \dontrun{} should only be used if the example really cannot be executed (e.g. because of missing additional software, missing API keys, ...) by
the user. That's why wrapping examples in \dontrun{} adds the comment ("# Not run:") as a warning for the user. Does not seem necessary. Please unwrap the examples if they are executable in < 5 sec, or replace \dontrun{} with \donttest{}.

Examples unwrapped and adapted to run in < 5 sec.

> You write information messages to the console that cannot be easily suppressed. It is more R like to generate objects that can be used to extract the
information a user is interested in, and then print() that object. Instead of print()/cat() rather use message()/warning()  or if(verbose)cat(..) (or maybe stop()) if you really have to write text to  the console. (except for print, summary, interactive functions)

`cat()` and `print()` to console are now only used in print functions (`print.iso_file_list`, `print.iso_file`, `print.binary_structure_map`), all other instances switched to `message()` or `warning()`.